backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.5.1
pydantic==2.8.2
pydantic-extra-types==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
apscheduler==3.10.4
email-validator==2.2.0
